refactor: Game 클래스 분리
